Refactor model file names
